test: StringArrayConverter.kt 이동
